Replace the usage of IdPreferenceManager by ConfigManager. Still need to do it in fingerprint and for the orchestrator
Former-commit-id: 87097f794141b032eaa9b40a77e93ed33fe5d4e0 [formerly 9fc7bb8ccff07427256a65bff37b1922f433be0b]
Former-commit-id: e0eb6f5c675ea80834c2193742cd8ba17379bed2

+++ face/src/main/java/com/simprints/face/models/FaceDetection.kt
@@ -74,6 +74,6 @@ data class FaceDetection(
         )
 
     fun hasValidStatus(): Boolean = status == Status.VALID || status == Status.VALID_CAPTURING
-    fun isAboveQualityThreshold(qualityThreshold: Float): Boolean =
+    fun isAboveQualityThreshold(qualityThreshold: Int): Boolean =
         face?.let { it.quality > qualityThreshold } ?: false
 }
